popular divs made into their own components to get usestate to work
