+++ src/App.jsx
@@ -4,14 +4,14 @@ import './App.css';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import Navigation from './Components/Navigation';
 import Slider from './Components/Sldier/Slider';
-import Popular from './Components/Popular/Popular';
+import PopularContainer from './Components/Popular/PopularContainer';
 
 function App() {
   return (
     <>
       <Navigation />
       <Slider />
-      <Popular />
+      <PopularContainer />
     </>
   );
 }
