+++ src/Components/Popular/Popular.css
@@ -2,3 +2,7 @@
   opacity: 0.5;
   transition: opacity 0.3s;
 }
+
+.custom-text-purple {
+  color: var(--secondary-color) !important;
+}
