+++ src/index.css
@@ -1,0 +1,8 @@
+:root {
+  --primary-color: #16151d;
+  --background-color: #312f40;
+  --secondary-bg-color: #23222e;
+  --black: #000;
+  --secondary-color: #9b51e0;
+  --white: #fff;
+}
